Set optimal data types for arrays

+++ MDP_EM/MDP_EM/policy_inference.py
@@ -50,7 +50,7 @@ class PolicyInference(object):
             self.mdp.policy = {state: {act: exp_Q[int(state)][act]/sum_exp_Q[int(state)] for act in self.mdp.action_list}
                                for state in self.mdp.states}
 
-    def gradientAscent(self, histories, theta_0=None, do_print=False, use_precomputed_phi=False):
+    def gradientAscent(self, histories, theta_0=None, do_print=False, use_precomputed_phi=False, dtype=np.float64):
         """
         @brief Performs Policy inference using Gradient Ascent from Section 7.2.1 of
         Sugiyama, 2015.
@@ -65,7 +65,7 @@ class PolicyInference(object):
         (num_episodes, num_steps) = self.histories.shape
         traj_samples = range(num_episodes) # Will be shuffled every iteration.
         # Precompute observed actions for all episodes.
-        observed_action_indeces = np.empty([num_episodes, num_steps], dtype=int)
+        observed_action_indeces = np.empty([num_episodes, num_steps], dtype=np.int8)
         for episode in xrange(num_episodes):
             for t_step in xrange(1, num_steps):
                 this_state = self.histories[episode, t_step-1]
@@ -76,7 +76,7 @@ class PolicyInference(object):
         # Initialize Weight vector, theta.
         if theta_0 == None:
             test_phi = self.mdp.phi(str(1), 'East')
-            theta_0 = np.empty([test_phi.size, 1]).T
+            theta_0 = np.empty([test_phi.size, 1], dtype=dtype).T
             for kern_idx in xrange(self.mdp.num_kern):
                 for act_idx, act in enumerate(self.mdp.action_list):
                         theta_0[0][kern_idx*self.mdp.num_actions+act_idx]= 1.0 / (theta_0.size)
@@ -86,7 +86,7 @@ class PolicyInference(object):
 
         # Velocity vector can be thought of as the momentum of the gradient descent. It is used to carry the theta
         # estimate through local minimums. https://wiseodd.github.io/techblog/2016/06/22/nn-optimization/.
-        velocity = np.zeros([theta_size])
+        velocity = np.zeros([theta_size], dtype=dtype)
         velocity_memory = 0.9
 
 
@@ -105,15 +105,15 @@ class PolicyInference(object):
         thresh = 0.05
         eps = 0.005
         iter_count = 0
-        inverse_temp = 1.0
+        inverse_temp = np.float16(1.0)
         # Larger value of inverse_temp_rate causes the temperature to cool faster, reduces oscilation. Set to 0 to
         # remove effect of temperature cooling.
-        inverse_temp_rate =  1.0
+        inverse_temp_rate =  np.float16(1.0)
         delta_theta_norm = np.inf
         theta = self.mdp.theta # Use a local reference to theta for look-up speed.
 
         # Initialize arrays for intermediate computations.
-        phis = np.zeros([self.mdp.num_states, self.mdp.num_actions, theta_size])
+        phis = np.zeros([self.mdp.num_states, self.mdp.num_actions, theta_size], dtype=dtype)
         for state in self.mdp.state_vec:
             for act_idx, act in enumerate(acts_list):
                 for kern_idx in xrange(theta_size):
@@ -131,7 +131,7 @@ class PolicyInference(object):
                 tic = time.clock()
             iter_count += 1
             inverse_temp += inverse_temp_rate
-            temp = 1.0 / inverse_temp
+            temp = np.float16(1.0) / inverse_temp
             prev_theta = deepcopy(theta)
             random.shuffle(traj_samples)
             traj_queue = deque(traj_samples)
